// frontend/src/api/generated/models/userAdminResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Manufacturing Information System API
 * API untuk aplikasi core MIS.
 * OpenAPI spec version: 1.0.0
 */
import type { DeletionEligibility } from './deletionEligibility.ts';
import type { MembershipAdminResponse } from './membershipAdminResponse.ts';

export interface UserAdminResponse {
  id: number;
  username: string;
  first_name: string;
  last_name: string;
  full_name: string;
  email: string;
  role: string;
  is_active: boolean;
  managed_by_server: boolean;
  editable: boolean;
  memberships: MembershipAdminResponse[];
  deletion_eligibility: DeletionEligibility;
}